# apps/mobile/api/generated/openapi.json
{
    "openapi": "3.0.1",
    "info": { "title": "OpenAPI definition", "version": "v0" },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/elo-simulation": {
            "get": {
                "tags": ["elo-simulation-controller"],
                "operationId": "getEloSimulation",
                "description": "A group's season computed by the leaderboard's Elo with other weights, every game's breakdown, and how well the ratings predict the next game over all seasons. Weights left out stay at the default; without seasonId the running season.",
                "parameters": [
                    {
                        "name": "inviteCode",
                        "in": "query",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "The group's invite code; it is all the simulator needs to open a group."
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "k",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "number", "format": "double" }
                    },
                    {
                        "name": "marginWeight",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "number", "format": "double" }
                    },
                    {
                        "name": "perPoint",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "number", "format": "double" }
                    },
                    {
                        "name": "topWeight",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "number", "format": "double" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeEloSimulationDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["elo-simulation-controller"],
                "operationId": "simulateEloWithTestGames",
                "description": "Like GET, with made-up test games counted where they say and the group's running live matches (teams as the app reduces them from their ops) counted as if they ended now. Nothing is stored. The standings then compare with the stored games alone.",
                "parameters": [
                    {
                        "name": "inviteCode",
                        "in": "query",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "The group's invite code; it is all the simulator needs to open a group."
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "k",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "number", "format": "double" }
                    },
                    {
                        "name": "marginWeight",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "number", "format": "double" }
                    },
                    {
                        "name": "perPoint",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "number", "format": "double" }
                    },
                    {
                        "name": "topWeight",
                        "in": "query",
                        "required": false,
                        "schema": { "type": "number", "format": "double" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeEloSimulationDto"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EloSimulationRequestDto"
                            }
                        }
                    },
                    "required": true
                }
            }
        },
        "/elo-simulation/search": {
            "get": {
                "tags": ["elo-simulation-controller"],
                "operationId": "searchEloWeights",
                "description": "Tries a grid of Elo weights on all of a group's seasons and returns the one that predicts its games best.",
                "parameters": [
                    {
                        "name": "inviteCode",
                        "in": "query",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "The group's invite code; it is all the simulator needs to open a group."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeEloSearchDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "tags": ["group-controller"],
                "operationId": "getGroupById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeGroupDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["group-controller"],
                "operationId": "updateGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GroupCreateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeGroupDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{id}/wallpaper": {
            "put": {
                "tags": ["group-controller"],
                "operationId": "setWallpaper",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssetCropDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeAssetUploadResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["group-controller"],
                "operationId": "deleteWallpaper",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeGroupDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{groupId}/seasons/{seasonId}/rules": {
            "get": {
                "tags": ["rule-controller"],
                "operationId": "getRules",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "seasonId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeListRuleDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["rule-controller"],
                "operationId": "writeRules",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "seasonId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/RuleCreateDto"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeListRuleDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{groupId}/seasons/{seasonId}/rule-moves/{ruleMoveId}": {
            "put": {
                "tags": ["rule-move-controller"],
                "operationId": "updateRuleMove",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "seasonId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "ruleMoveId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RuleMoveCreateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeRuleMoveDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{groupId}/seasons/{seasonId}/matches/{id}": {
            "get": {
                "tags": ["match-controller"],
                "operationId": "getMatchById",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "seasonId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeMatchDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["match-controller"],
                "operationId": "updateMatch",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "seasonId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatchCreateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeMatchDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["match-controller"],
                "operationId": "deleteMatchById",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "seasonId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeString"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{groupId}/seasons/{seasonId}/matches/{id}/photos/{teamId}": {
            "put": {
                "tags": ["match-controller"],
                "operationId": "setPhoto",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "teamId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "seasonId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeAssetUploadResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["match-controller"],
                "operationId": "deletePhoto",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "teamId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name": "seasonId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeTeamDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/groups/{groupId}/active-season": {
            "put": {
                "tags": ["season-controller"],
                "operationId": "startNewSeason",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeasonCreateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeSeasonDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/groups/{id}/leave": {
            "post": {
                "tags": ["group-controller"],
                "operationId": "leaveGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseEnvelopeString"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EloSimulationRequestDto": {
                "type": "object",
                "properties": {
                    "testGames": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EloTestGameDto"
                        }
                    },
                    "liveMatches": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EloLiveMatchDto"
                        }
                    }
                },
                "required": []
            },
...
            "ResponseEnvelopeEloSimulationDto": {
                "type": "object",
                "properties": {
                    "status": { "type": "string", "enum": ["OK", "ERROR"] },
                    "httpCode": { "type": "integer", "format": "int32" },
                    "data": { "$ref": "#/components/schemas/EloSimulationDto" },
                    "error": { "$ref": "#/components/schemas/ErrorDetails" }
                }
            },
            "ResponseEnvelopeEloSearchDto": {
                "type": "object",
                "properties": {
                    "status": { "type": "string", "enum": ["OK", "ERROR"] },
                    "httpCode": { "type": "integer", "format": "int32" },
                    "data": { "$ref": "#/components/schemas/EloSearchDto" },
                    "error": { "$ref": "#/components/schemas/ErrorDetails" }
                }
            },
...
        }
    }
}